requirements.txt
Flask==2.3.2
openai==0.28.0
python-dotenv==1.0.0
nltk==3.8.1
flask_socketio==5.3.5
eventlet==0.30.2
# gevent-websocket==0.10.1
# gunicorn==21.2.0